requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
ultralytics>=8.1.0
pillow>=10.2.0
numpy>=1.24.0
python-multipart>=0.0.9





